chore(Profitwell): move sdk loader to different file
