+++ src/integrations/ProfitWell/browser.js
@@ -1,8 +1,8 @@
 /* eslint-disable class-methods-use-this */
 import get from 'get-value';
 import logger from '../../utils/logUtil';
-import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
 import { NAME } from './constants';
+import { loader } from './loader';
 
 class ProfitWell {
   constructor(config, analytics, destinationInfo) {
@@ -33,19 +33,13 @@ class ProfitWell {
     scriptTag.setAttribute('data-pw-auth', window.publicApiKey);
     document.body.appendChild(scriptTag);
 
-    (function (i, s, o, g, r, a, m) {
-      i[o] =
-        i[o] ||
-        function () {
-          (i[o].q = i[o].q || []).push(arguments);
-        };
-      a = s.createElement(g);
-      m = s.getElementsByTagName(g)[0];
-      a.async = 1;
-      a.setAttribute('data-loader', LOAD_ORIGIN);
-      a.src = `${r}?auth=${window.publicApiKey}`;
-      m.parentNode.insertBefore(a, m);
-    })(window, document, 'profitwell', 'script', 'https://public.profitwell.com/js/profitwell.js');
+    loader(
+      window,
+      document,
+      'profitwell',
+      'script',
+      'https://public.profitwell.com/js/profitwell.js',
+    );
   }
 
   isLoaded() {
